Adding mapping

+++ src/main/java/com/ubi/academicapplication/repository/ClassRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ubi.academicapplication.entity.ClassDetail;
 
 
-public interface ClassRepository extends JpaRepository<ClassDetail, Long> {
+public interface ClassRepository extends JpaRepository<ClassDetail, Integer> {
 
 
 }
